+++ backend/src/main/java/zipgo/review/domain/Review.java
@@ -8,7 +8,7 @@ import zipgo.auth.exception.AuthException;
 import zipgo.common.entity.BaseTimeEntity;
 import zipgo.member.domain.Member;
 import zipgo.petfood.domain.PetFood;
-import zipgo.review.domain.type.AdverseReactionName;
+import zipgo.review.domain.type.AdverseReactionType;
 import zipgo.review.domain.type.StoolCondition;
 import zipgo.review.domain.type.TastePreference;
 
@@ -62,7 +62,7 @@ public class Review extends BaseTimeEntity {
 
     public void addAdverseReactions(List<String> adverseReactionNames) {
         List<AdverseReaction> adverseReactions = adverseReactionNames.stream()
-                .map(name -> new AdverseReaction(AdverseReactionName.from(name)))
+                .map(name -> new AdverseReaction(AdverseReactionType.from(name)))
                 .toList();
 
         for (AdverseReaction adverseReaction : adverseReactions) {
