+++ backend/src/test/java/zipgo/review/application/ReviewQueryServiceTest.java
@@ -18,7 +18,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static zipgo.brand.domain.fixture.BrandFixture.식품_브랜드_생성하기;
 import static zipgo.petfood.domain.fixture.PetFoodFixture.키워드_없이_식품_초기화;
-import static zipgo.review.domain.type.AdverseReactionName.NONE;
+import static zipgo.review.domain.type.AdverseReactionType.NONE;
 import static zipgo.review.domain.type.StoolCondition.SOFT_MOIST;
 import static zipgo.review.domain.type.TastePreference.EATS_VERY_WELL;
 import static zipgo.review.fixture.AdverseReactionFixture.눈물_이상반응;
@@ -51,7 +51,7 @@ class ReviewQueryServiceTest extends QueryServiceTest {
         Member 멤버 = memberRepository.save(무민());
         petFoodRepository.save(식품);
         Review 극찬_리뷰 = reviewRepository.save(극찬_리뷰_생성(멤버, 식품));
-        Review 혹평_리뷰_생성 = 혹평_리뷰_생성(멤버, 식품, List.of(눈물_이상반응().getAdverseReactionName().getDescription(), 먹고_토_이상반응().getAdverseReactionName().getDescription()));
+        Review 혹평_리뷰_생성 = 혹평_리뷰_생성(멤버, 식품, List.of(눈물_이상반응().getAdverseReactionType().getDescription(), 먹고_토_이상반응().getAdverseReactionType().getDescription()));
         reviewRepository.save(혹평_리뷰_생성);
 
         //when
@@ -66,7 +66,7 @@ class ReviewQueryServiceTest extends QueryServiceTest {
                 () -> assertThat(찾은_리뷰.getComment()).isEqualTo("우리 아이랑 너무 잘 맞아요!"),
                 () -> assertThat(찾은_리뷰.getTastePreference()).isEqualTo(EATS_VERY_WELL),
                 () -> assertThat(찾은_리뷰.getStoolCondition()).isEqualTo(SOFT_MOIST),
-                () -> assertThat(찾은_리뷰.getAdverseReactions().get(0).getAdverseReactionName()).isEqualTo(NONE)
+                () -> assertThat(찾은_리뷰.getAdverseReactions().get(0).getAdverseReactionType()).isEqualTo(NONE)
         );
     }
