+++ backend/src/test/resources/data.sql
@@ -27,7 +27,7 @@ values (1, 1, 5, '우리 아이랑 너무 잘 맞아요!', 'EATS_VERY_WELL', 'SO
 insert into review(member_id, pet_food_id, rating, comment, taste_preference, stool_condition, created_at)
 values (1, 1, 1, '우리 아이가 한 입 먹고 더 안 먹어요 ㅡ.ㅡ 책임지세요.', 'NOT_AT_ALL', 'DIARRHEA', '2023-07-28');
 
-insert into adverse_reaction(adverse_reaction_name, review_id)
+insert into adverse_reaction(adverse_reaction_type, review_id)
 values ('TEARS', 2);
-insert into adverse_reaction(adverse_reaction_name, review_id)
+insert into adverse_reaction(adverse_reaction_type, review_id)
 values ('VOMITING', 2);
